% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit_subjects.R
\name{baseline_intensities_from_coxmod}
\alias{baseline_intensities_from_coxmod}
\title{Recover baseline intensities (in the form of intensity_matrices) from a 
coxph() fit on multi-state data.}
\usage{
baseline_intensities_from_coxmod(object, tmat)
}
\arguments{
\item{object}{A \code{\link[survival:coxph]{coxph}} object describing the
fit of the multi-state model, must contain a 'strata()' term.}

\item{tmat}{A transition matrix as created by \code{\link[mstate:transMat]{transMat}}.}
}
\description{
Recover baseline intensities (in the form of intensity_matrices) from a 
coxph() fit on multi-state data.
}
\author{
Daniel Gomon
}
\keyword{internal}
