% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.coxph.relsurv.R
\name{coxph.relsurv}
\alias{coxph.relsurv}
\title{Fit Extended Proportional Hazards Regression Model with Relative Survival}
\usage{
coxph.relsurv(
  formula,
  data,
  na.action,
  split.transitions,
  ratetable = relsurv::slopop,
  time.format = "days",
  rmap,
  init,
  bwin,
  centered,
  cause,
  ...
)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function}

\item{data}{The data used for fitting the model}

\item{na.action}{A missing-data filter function. This is applied to the model.frame after any subset argument has been used. Default is options()$na.action}

\item{split.transitions}{An integer vector containing the numbered transitions that should be split. Use same numbering as in the given transition matrix}

\item{ratetable}{The population mortality table. A table of event rates, organized as a ratetable object, see for example relsurv::slopop. Default is slopop}

\item{time.format}{Define the time format which is used in the data. Possible options: c('days', 'years', 'months'). Default is 'days'}

\item{rmap}{An optional list to be used if the variables in the data are not organized (and named) in the same way as in the ratetable object}

\item{init}{Vector of initial values of the iteration used when estimating the effects for the excess hazards. Default initial value is zero for all variables}

\item{bwin}{Controls the bandwidth used for smoothing in the EM algorithm when estimating the effects for the excess hazards. The follow-up time is divided into quartiles and bwin specifies a factor by which the maximum between events time length on each interval is multiplied. The default bwin=-1 lets the function find an appropriate value. If bwin=0, no smoothing is applied}

\item{centered}{If TRUE, all variables are centered before fitting the EM algorithmg for the excess hazards and the baseline excess hazard is calculated accordingly. Default is FALSE}

\item{cause}{An optional vector of the same length as the number of cases if cause of death is (partially) given. Use 0 for population deaths, 1 for disease-specific deaths, 2 (default) for unknown.}

\item{...}{Other arguments that will be passed to coxph}
}
\value{
Returns a coxph.relsurv object that contains the coefficient estimates for the extended model
}
\description{
An extension of the survival::coxph function where one can split death-related transitions
and model the corresponding excess hazards using relative survival.
}
\seealso{
\code{\link{coxph}}, \code{\link{msfit.relsurv}}, \code{\link[relsurv]{rsadd}}
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
