% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EBMT year of relapse data}
\alias{EBMT year of relapse data}
\alias{ebmt1}
\title{Data from the European Society for Blood and Marrow Transplantation (EBMT)}
\format{
A data frame, see \code{\link{data.frame}}.
}
\source{
We acknowledge the European Society for Blood and Marrow
Transplantation (EBMT) for making available these data. Disclaimer: these
data were simplified for the purpose of illustration of the analysis of
competing risks and multi-state models and do not reflect any real life
situation. No clinical conclusions should be drawn from these data.
}
\description{
A data frame of 1977 patients transplanted for CML. The included variables
are \describe{ \item{patid}{Patient identification number} \item{srv}{Time
in days from transplantation to death or last follow-up}
\item{srvstat}{Survival status; 1 = death; 0 = censored} \item{rel}{Time in
days from transplantation to relapse or last follow-up}
\item{relstat}{Relapse status; 1 = relapsed; 0 = censored}
\item{yrel}{Calendar year of relapse; factor with levels "1993-1996","
1997-1999", "2000-"} \item{age}{Patient age at transplant (years)}
\item{score}{Gratwohl score; factor with levels "Low risk", "Medium risk",
"High risk"} }
}
\keyword{datasets}
