% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etm2msdata.R
\name{msdata2etm}
\alias{msdata2etm}
\title{msdata to etm format}
\usage{
msdata2etm(msdata, id, covs)
}
\arguments{
\item{msdata}{Multi-state data in \code{msdata} format, as used in
\code{mstate}}

\item{id}{Column name identifying the subject id}

\item{covs}{Vector of column names containing covariates to be included}
}
\description{
msdata to etm format
}
