% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit_subjects.R
\name{msfit_subjects}
\alias{msfit_subjects}
\title{Compute subject-specific transition hazards for multiple subjects at once}
\usage{
msfit_subjects(object, newdata, trans)
}
\arguments{
\item{object}{A \code{\link[survival:coxph]{coxph}} object describing the
fit of the multi-state model, must contain a 'strata()' term.}

\item{newdata}{A \code{data.frame} containing K rows per
subject in the data, with K the number of transitions possible in the model.
The following named columns must be present:
\describe{
  \item{\code{id}:}{Unique identifier of the subject this row relates to, must be numeric or character;}
  \item{\code{trans}:}{Transition number in the \code{'transMat'} \code{trans} 
  this row relates to;}
  \item{\code{"variables"}:}{The subject-specific covariates (appearing in the \code{coxph} formula),
  ordered according to transition number in \code{trans}.}
} Note that newdata must contain a column containing the variable which was 
used to determine the stratum of a transition in \code{object} (usually this is 'trans').}

\item{trans}{Transition matrix describing the states and transitions in the
multi-state model. See \code{trans} in \code{\link{msprep}} for more
detailed information}
}
\value{
An object of class \code{"msfit.subjects"}, which is a list containing
\item{intensities}{A 4-dimensional \code{array} containing the estimated cumulative 
intensity increments at \code{times}. The first and second dimensions pertain 
to the to/from states of the transitions, the third to the \code{times} and 
the fourth to the unique subject \code{ids}.} 
\item{times}{The unique times at which the intensity increments were determined.}
\item{ids}{The unique id's used to represent the individual subjects, taken 
from the 'id' column in \code{newdata}.} 
\item{trans}{The transition matrix used.}
Looking for an 'msfit' object for a single subject? See 
\code{\link{msfit_subjects_to_msfit}}.
}
\description{
This function computes subject-specific or overall cumulative transition
hazards for each of the possible transitions in the multi-state model.
Contrary to \code{\link{msfit}}, this function allows to 
calculate hazards for multiple subjects at once, but does not allow
to calculate (co)variances (yet).
}
\details{
The data frame needs to have one row for each transition in the multi-state
model, per subject: (n_subjects x n_transitions) rows in total.
Contrary to \code{\link{msfit}}, it is not necessary to 
manually specify the \code{strata} variable,
as long as the strata can be determined from the data using the formula 
used in \code{object}. For details refer to de Wreede, Fiocco & Putter (2010). 
So far,
the results have been checked only for the \code{"breslow"} method of
dealing with ties in \code{\link[survival:coxph]{coxph}}, so this is
recommended.
The goal of this function is to determine the subject specific intensity matrices
\eqn{dA_i(t)}{dA_i(t)}, which can be further used to determine transition probabilities
through the product integral \deqn{P_i(s,t) = \prod_{(s,t]}(I + dA_i(u))}{P_i(s,t) = \prod_{(s,t])}(I + dA_i(u))}
}
\examples{

# transition matrix for illness-death model
tmat <- trans.illdeath()
# data in wide format, for transition 1 this is dataset E1 of
# Therneau & Grambsch (2000)
tg <- data.frame(illt=c(1,1,6,6,8,9),ills=c(1,0,1,1,0,1),
        dt=c(5,1,9,7,8,12),ds=c(1,1,1,1,1,1),
        x1=c(1,1,1,0,0,0),x2=c(6:1))
# data in long format using msprep
tglong <- msprep(time=c(NA,"illt","dt"),status=c(NA,"ills","ds"),
		data=tg,keep=c("x1","x2"),trans=tmat)
# expanded covariates
tglong <- expand.covs(tglong,c("x1","x2"))
# Cox model with different covariate
cx <- coxph(Surv(Tstart,Tstop,status)~x1.1+x2.2+strata(trans),
	data=tglong,method="breslow")
#Fit on multiple subjects at once, by providing 'id' column.
#cx was fit using strata(trans), so only having a 'trans' column suffices
newdata <- data.frame(id=rep(1:3, each = 3),x1.1=c(0,0,0,1,0,1,0,1,0),
                      x2.2=c(0,1,0,0,0,0,1,0,1), trans = rep(1:3, 3))
msf_subj <- msfit_subjects(cx,newdata,trans=tmat)
#Extract an 'msfit' object for subject 1
msf_subj1 <- msfit_subjects_to_msfit(msf_subj, 1)
#Can now use standard 'msfit' plotting capabilities
plot(msf_subj1)

}
\references{
Putter H, Fiocco M, Geskus RB (2007). Tutorial in biostatistics:
Competing risks and multi-state models. \emph{Statistics in Medicine}
\bold{26}, 2389--2430.

Therneau TM, Grambsch PM (2000). \emph{Modeling Survival Data: Extending the
Cox Model}. Springer, New York.

de Wreede LC, Fiocco M, and Putter H (2010). The mstate package for
estimation and prediction in non- and semi-parametric multi-state and
competing risks models. \emph{Computer Methods and Programs in Biomedicine}
\bold{99}, 261--274.

de Wreede LC, Fiocco M, and Putter H (2011). mstate: An R Package for the
Analysis of Competing Risks and Multi-State Models. \emph{Journal of
Statistical Software}, Volume 38, Issue 7.
}
\seealso{
\code{\link{plot.msfit.subjects}}, \code{\link{msfit_subjects_to_msfit}}
}
\author{
Daniel Gomon
}
\keyword{survival}
