% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit_subjects.R
\name{msfit_subjects_to_msfit}
\alias{msfit_subjects_to_msfit}
\title{Obtain an 'msfit' object for a single subject from an 'msfit.subjects' 
object.}
\usage{
msfit_subjects_to_msfit(object, id)
}
\arguments{
\item{object}{A \code{'msfit.subjects'} object.}

\item{id}{A single subject_id, must be contained in \code{object$ids}.}
}
\value{
An \code{\link[mstate:msfit]{msfit}} object for a single subject.
}
\description{
Obtain an 'msfit' object for a single subject from an 'msfit.subjects' 
object.
}
\seealso{
\code{\link{msfit}}
}
\author{
Daniel Gomon
}
