\name{mstate-package}
\alias{mstate-package}
\alias{mstate}
\docType{package}
\title{
Data preparation, estimation and prediction in multi-state models
}
\description{
Functions for data preparation, descriptives,
(hazard) estimation and prediction (Aalen-Johansen) in competing
risks and multi-state models.}
\details{
\tabular{ll}{
Package: \tab mstate\cr
Type: \tab Package\cr
Version: \tab 0.2.10\cr
Date: \tab 2016-12-03\cr
License: \tab GPL 2.0\cr
}
}
\author{
Liesbeth de Wreede, Marta Fiocco, Hein Putter.
Maintainer: Hein Putter <H.Putter@lumc.nl>
}
\references{
Putter H, Fiocco M, Geskus RB (2007). Tutorial in biostatistics:
Competing risks and multi-state models. \emph{Statistics in Medicine} \bold{26},
2389--2430.

de Wreede LC, Fiocco M, and Putter H (2010). The mstate package
for estimation and prediction in non- and semi-parametric multi-state
and competing risks models. \emph{Computer Methods and Programs in
Biomedicine} \bold{99}, 261--274.

de Wreede LC, Fiocco M, and Putter H (2011). mstate: An R Package
for the Analysis of Competing Risks and Multi-State Models. \emph{Journal of
Statistical Software}, Volume 38, Issue 7.
}
\keyword{ package }
