% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Cuminc.R
\name{plot.Cuminc}
\alias{plot.Cuminc}
\title{Plot method for cuminc objects}
\usage{
\method{plot}{Cuminc}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"Cuminc"} to be printed or plotted}

\item{\dots}{Further arguments to plot or print method}
}
\description{
Plot method for cuminc objects
}
