% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.coxph.relsurv.R
\name{print.coxph.relsurv}
\alias{print.coxph.relsurv}
\title{Print method for coxph.relsurv object}
\usage{
\method{print}{coxph.relsurv}(x, digits = max(1L, getOption("digits") - 3L), signif.stars = FALSE, ...)
}
\arguments{
\item{x}{Object of class coxph.relsurv to be printed}

\item{digits}{Number of digits}

\item{signif.stars}{Should significance stars be printed}

\item{...}{Further arguments to print}
}
\description{
Print method for coxph.relsurv object
}
\seealso{
\code{\link{coxph.relsurv}}
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
