% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.summary.coxph.relsurv.R
\name{print.summary.coxph.relsurv}
\alias{print.summary.coxph.relsurv}
\title{Print method for summary.coxph.relsurv objects}
\usage{
\method{print}{summary.coxph.relsurv}(x, digits = max(getOption("digits") - 3, 3), expand = FALSE, ...)
}
\arguments{
\item{x}{The result of a call to summary.coxph.relsurv}

\item{digits}{significant digits to print}

\item{expand}{If the summary is for a multi-state coxph fit, print the results in an expanded format}

\item{...}{For future methods}
}
\description{
Produces a printed summary of a fitted coxph.relsurv model
}
\seealso{
\code{\link{coxph.relsurv}}, \code{\link{summary.coxph.relsurv}}
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
