% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.msfit.R
\name{print.summary.msfit}
\alias{print.summary.msfit}
\title{Print method for summary.msfit object}
\usage{
\method{print}{summary.msfit}(x, complete = FALSE, ...)
}
\arguments{
\item{x}{Object of class 'summary.msfit', to be printed}

\item{complete}{Whether or not the complete estimated cumulative transition
intensities should be printed (\code{TRUE}) or not (\code{FALSE}); default
is \code{FALSE}, in which case the estimated cumulative transition hazards
will be printed for the first and last 6 time points of each transition or
of the selected times (or all when there are at most 12 of these time points}

\item{\dots}{Further arguments to print}
}
\description{
Print method for summary.msfit object
}
\examples{
\dontrun{
# If all time points should be printed, specify complete=TRUE in the print statement
print(x, complete=TRUE)
}

}
