% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.helpers.R
\name{set_colours}
\alias{set_colours}
\title{Automatic choice of default colours for plot}
\usage{
set_colours(n, type)
}
\arguments{
\item{n}{Number of groups/lines/areas to be plotted. Generally number
of states or transitions}

\item{type}{Either "line" or "areas"}
}
\value{
Vector of colours
}
\description{
Automatic choice of default colours for plot
}
