% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit_subjects.R
\name{subject_specific_intensity_matrices}
\alias{subject_specific_intensity_matrices}
\title{Calculate the subject specific intensity matrices}
\usage{
subject_specific_intensity_matrices(
  subject_specific_risks,
  baseline_intensities,
  trans
)
}
\description{
For each subject, calculate a 3D array containing the states (from, to) in the
first two dimension and the times in the third.
This is then stored in a 4D array, with the 4th dimension indicating each unique subject.
}
\author{
Daniel Gomon
}
\keyword{internal}
