% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.summary.coxph.relsurv.R
\name{summary.coxph.relsurv}
\alias{summary.coxph.relsurv}
\title{Summary method for coxph.relsurv object}
\usage{
\method{summary}{coxph.relsurv}(object, conf.int = 0.95, scale = 1, ...)
}
\arguments{
\item{object}{The result of a coxph.relsurv fit}

\item{conf.int}{Level for computation of the confidence intervals. If set to FALSE no confidence intervals are printed}

\item{scale}{Vector of scale factors for the coefficients, defaults to 1. The printed coefficients, se, and confidence intervals will be associated with one scale unit}

\item{...}{For future methods}
}
\value{
An object of class summary.coxph.relsurv
}
\description{
Produces a summary of a fitted coxph.relsurv model.
}
\seealso{
\code{\link{coxph.relsurv}}
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
