% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Cuminc.R
\name{summary.Cuminc}
\alias{summary.Cuminc}
\title{Summary method for a summary.Cuminc object}
\usage{
\method{summary}{Cuminc}(object, ...)
}
\arguments{
\item{object}{Object of class 'Cuminc', to be summarised}

\item{\dots}{Further arguments to summarise}
}
\description{
Summary method for a summary.Cuminc object
}
