% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etm2msdata.R
\name{trans2tra}
\alias{trans2tra}
\title{Convert transition matrix from mstate to etm format}
\usage{
trans2tra(trans)
}
\arguments{
\item{trans}{Transition matrix in \code{mstate} format}
}
\description{
Convert transition matrix from mstate to etm format
}
