% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit_subjects.R
\name{trans_specific_risks}
\alias{trans_specific_risks}
\title{Calculate subject specific risks for subjects in newdata}
\usage{
trans_specific_risks(object, newdata, trans, n_subjects, subject_ids)
}
\arguments{
\item{object}{A \code{\link[survival:coxph]{coxph}} object describing the
fit of the multi-state model, must contain a 'strata()' term.}

\item{newdata}{A \code{data.frame} containing K rows per
subject in the data, with K the number of transitions possible in the model.
The following named columns must be present:
\describe{
  \item{\code{id}:}{Unique identifier of the subject this row relates to, must be numeric or character;}
  \item{\code{trans}:}{Transition number in the \code{'transMat'} \code{trans} 
  this row relates to;}
  \item{\code{"variables"}:}{The subject-specific covariates (appearing in the \code{coxph} formula),
  ordered according to transition number in \code{trans}.}
} Note that newdata must contain a column containing the variable which was 
used to determine the stratum of a transition in \code{object} (usually this is 'trans').}

\item{trans}{Transition matrix describing the states and transitions in the
multi-state model. See \code{trans} in \code{\link{msprep}} for more
detailed information}
}
\description{
Return a 2 dimensional array, with subjects in the first dimension and
transition (numbers) in the second. Can expand later to include time-dependent
covariates by introducing extra dimension. Entries of the matrix are exp(lp)_i^m,
with i denoting the subject and m the transition.
}
\author{
Daniel Gomon
}
\keyword{internal}
