% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.varHaz.fixed.R
\name{varHaz.fixed}
\alias{varHaz.fixed}
\title{Upgrade the varHaz object}
\usage{
varHaz.fixed(varHaz, link_trans, varHaz_original)
}
\arguments{
\item{varHaz}{The varHaz object (present in a msfit object).}

\item{link_trans}{A list that gives the linkage between the original and upgraded 
transition matrix.}

\item{varHaz_original}{The original varHaz object from msfit (without the eventual time conversion).}
}
\value{
Return the upgraded varHaz object containing variances for the split transitions.
}
\description{
A function that upgrades varHaz from the msfit object where the variances are
estimated using the Greenwood estimator; it is further assumed that variances for the population
hazards are equal to zero.
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
