% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.mirror.pt.R
\name{vis.mirror.pt}
\alias{vis.mirror.pt}
\title{Mirror plot comparing two probtrans objects}
\usage{
vis.mirror.pt(
  x,
  titles,
  size_titles = 15,
  breaks_x_left,
  breaks_x_right,
  ylim,
  xlab = "Time",
  ylab = "Probability",
  legend.pos = "right"
)
}
\arguments{
\item{x}{A list of two plots as returned by plot.probtrans with
use.ggplot = T. The first element will be on the left of the mirror plot,
 and the second on the right}

\item{titles}{A character vector c("Title for left", "Title for right")}

\item{size_titles}{Numeric, size of the title text}

\item{breaks_x_left}{Numeric vector specifying axis breaks on the left plot}

\item{breaks_x_right}{Numeric vector specifying axis breaks on the right plot}

\item{ylim}{Numeric vector, limits of the y-axis. Default is c(0, 1)}

\item{xlab}{A title for the x-axis, default is "Time"}

\item{ylab}{A title for the y-axis, default is "Probability"}

\item{legend.pos}{Position of the legend, default is "right"}
}
\value{

}
\description{
A mirror plot for comparing two different probtrans objects. Useful
for comparing predicted probabilities for different levels of a covariate,
or for different subgroups.
}
