% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit_subjects.R
\name{baseline_intensities_from_coxmod}
\alias{baseline_intensities_from_coxmod}
\title{Recover baseline intensities (in the form of intensity_matrices) from a 
coxph() fit on multi-state data.}
\usage{
baseline_intensities_from_coxmod(object, tmat, newdata)
}
\arguments{
\item{object}{A \code{\link[survival:coxph]{coxph}} object describing the
fit of the multi-state model, must contain a 'strata()' term.}

\item{tmat}{A transition matrix as created by \code{\link[mstate:transMat]{transMat}}.}

\item{newdata}{A \code{data.frame} containing K rows per
subject in the data, with K the number of transitions possible in the model.
The following named columns must be present:
\describe{
  \item{\code{id}:}{Unique identifier of the subject this row relates to, must be numeric or character;}
  \item{\code{trans}:}{Transition number in the \code{'transMat'} \code{trans} 
  this row relates to;}
  \item{\code{"variables"}:}{The subject-specific covariates (appearing in the \code{coxph} formula),
  ordered according to transition number in \code{trans}.}
} Note that newdata must contain a column containing the variable which was 
used to determine the stratum of a transition in \code{object} (usually this is 'trans').}
}
\description{
Recover baseline intensities (in the form of intensity_matrices) from a 
coxph() fit on multi-state data.
}
\author{
Daniel Gomon
}
\keyword{internal}
