% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.msfit.coxph.relsurv.R
\name{msfit.coxph.relsurv}
\alias{msfit.coxph.relsurv}
\title{Compute subject-specific transition hazards with (co-)variances from coxph.relsurv}
\usage{
msfit.coxph.relsurv(
  object,
  newdata,
  variance = FALSE,
  vartype = c("aalen", "greenwood"),
  trans
)
}
\arguments{
\item{object}{A coxph.relsurv object}

\item{newdata}{A data frame with the same variable names as those that appear in the \code{coxph.relsurv} formula}

\item{variance}{Submitted to msfit for the overall hazards}

\item{vartype}{Submitted to msfit for the overall hazards}

\item{trans}{Transition matrix describing the states and transitions in the
multi-state model. See \code{trans} in \code{\link{msprep}} for more
detailed information}
}
\value{
An object of class \code{"msfit"}, which is a list containing
\item{Haz }{A data frame with \code{time}, \code{Haz}, \code{trans},
containing the estimated subject-specific hazards for each of the
transitions in the multi-state model} \item{varHaz }{A data frame with
\code{time}, \code{Haz}, \code{trans1}, \code{trans2} containing the
variances (\code{trans1}=\code{trans2}) and covariances
(\code{trans1}<\code{trans2}) of the estimated hazards. This element is only
returned when \code{variance}=\code{TRUE}} \item{trans}{The extended transition
matrix}
}
\description{
An extension of the msfit function for coxph.relsurv
}
\seealso{
\code{\link{coxph.relsurv}}, \code{\link{msfit.relsurv}}, \code{\link[relsurv]{rsadd}}
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
