% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsurv.msfit.relsurv.R
\name{msfit.relsurv}
\alias{msfit.relsurv}
\title{Extend a multi-state model using relative survival}
\usage{
msfit.relsurv(
  msfit.obj,
  data,
  split.transitions,
  ratetable = relsurv::slopop,
  rmap,
  time.format = "days",
  var.pop.haz = c("fixed", "bootstrap", "both"),
  B = 100,
  seed = NULL,
  add.times,
  substitution = TRUE,
  link_trans_ind = FALSE
)
}
\arguments{
\item{msfit.obj}{The msfit object which has to be upgraded}

\item{data}{The dataset used for fitting the msfit model}

\item{split.transitions}{An integer vector containing the numbered transitions that should be split. Use same numbering as in the given transition matrix}

\item{ratetable}{The population mortality table. A table of event rates, organized as a ratetable object, see for example relsurv::slopop. Default is slopop}

\item{rmap}{An optional list to be used if the variables in the dataset are not organized (and named) in the same way as in the ratetable object}

\item{time.format}{Define the time format which is used in the dataset Possible options: c('days', 'years', 'months'). Default is 'days'}

\item{var.pop.haz}{If 'fixed' (default), one assumes that the variance of the population hazards is zero. If 'bootstrap', one gets boostrapped variances for all transitions.}

\item{B}{Number of bootstrap replications. Relevant only if var.pop.haz == 'bootstrap'}

\item{seed}{Set seed}

\item{add.times}{Additional times at which hazards should be evaluated}

\item{substitution}{Whether function substitute should be used for rmap argument}

\item{link_trans_ind}{Choose whether the linkage between the old and new transition matrix should be saved. Default is FALSE.}
}
\value{
Returns a msfit object that now contains hazards and their variances (if provided) 
for the split (population and excess) transitions.
}
\description{
A function that takes a fitted msfit object and upgrades
it using relative survival, where chosen transitions are
split in population and excess transitions. This upgraded 
msfit object contains the split hazards based on the transition
matrix (transMat). The (co)variance matrix is also upgraded, if provided.
}
\seealso{
\code{\link{msfit}}
}
\author{
Damjan Manevski \email{damjan.manevski@mf.uni-lj.si}
}
