\name{mstate-package}
\alias{mstate-package}
\alias{mstate}
\docType{package}
\title{
Data preparation, estimation and prediction in multi-state models
}
\description{
Functions for data preparation, descriptives,
(hazard) estimation and prediction (Aalen-Johansen) in competing
risks and multi-state models.}
\details{
\tabular{ll}{
Package: \tab mstate\cr
Type: \tab Package\cr
Version: \tab 0.2.2\cr
Date: \tab 2009-09-15\cr
License: \tab GPL 2.0\cr
}
An overview of how to use the package, including the most important
functions.
}
\author{
Liesbeth de Wreede, Marta Fiocco, Hein Putter.
Maintainer: Hein Putter <H.Putter@lumc.nl>
}
\references{
Putter H, Fiocco M, Geskus RB (2007). Tutorial in biostatistics:
Competing risks and multi-state models. Statistics in Medicine 26,
2389--2430.

L. C. de Wreede, M. Fiocco, and H. Putter (2009). The mstate package
for estimation and prediction in non- and semi-parametric multi-state
and competing risks models. Submitted.
}
\keyword{ package }
