% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.mirror.pt.R
\name{vis.mirror.pt}
\alias{vis.mirror.pt}
\title{Mirror plot comparing two probtrans objects}
\usage{
vis.mirror.pt(
  x,
  titles,
  size_titles = 5,
  breaks_x_left,
  breaks_x_right,
  from = 1,
  cols,
  ord,
  xlab = "Time",
  ylab = "Probability",
  legend.pos = "right"
)
}
\arguments{
\item{x}{A list of two pt objects. The first element will be on the left of 
the mirror plot, and the second on the right}

\item{titles}{A character vector c("Title for left", "Title for right")}

\item{size_titles}{Numeric, size of the title text}

\item{breaks_x_left}{Numeric vector specifying axis breaks on the left plot}

\item{breaks_x_right}{Numeric vector specifying axis breaks on the right plot}

\item{from}{The starting state from which the probabilities are used to plot}

\item{cols}{A vector specifying colors for the different transitions;
default is a palette from green to red, when type=\code{"filled"} (reordered
according to \code{ord}, and 1 (black), otherwise}

\item{ord}{A vector of length equal to the number of states, specifying the
order of plotting in case type=\code{"stacked"} or \code{"filled"}}

\item{xlab}{A title for the x-axis, default is "Time"}

\item{ylab}{A title for the y-axis, default is "Probability"}

\item{legend.pos}{Position of the legend, default is "right"}
}
\value{
A ggplot2 object with the mirrored plot.
}
\description{
A mirror plot for comparing two different probtrans objects. Useful
for comparing predicted probabilities for different levels of a covariate,
or for different subgroups.
}
