% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.multiple.pt.R
\name{vis.multiple.pt}
\alias{vis.multiple.pt}
\title{Visualise multiple probtrans objects}
\usage{
vis.multiple.pt(x, to, labels, legend.title)
}
\arguments{
\item{x}{A list of plots as returned by plot(pt, use_ggplot = T)}

\item{to}{Character; destination state}

\item{labels}{Character vector labelling each element of x (e.g. label
for a reference patient) - so labels = c("Patient 1", "Patient 2")}
}
\value{
A plot visualising
}
\description{
Description here.
}
